+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 CHANGELOG
 =========
 
+## 1.0.5
+
+### Add `Fenom::AUTO_ESCAPE` support (feature #2)
+### Update documentation
+
+## 1.0.4 (2013-06-27)
+
+### Add nested level for {extends} and {use}
+### Small bug fix
+### Update documentation
+
 ## 1.0.3 (2013-06-20)
 
 ### Allow any callable for modifier (instead string)